app components switch: rename from toggle to switch
